added demo.js still not functional
